+++ src/dmmf/transformer.ts
@@ -1,9 +1,9 @@
 import { DMMF } from '@prisma/client/runtime'
+import { paginationStrategies, PaginationStrategy } from '../pagination'
 import { GlobalComputedInputs, GlobalMutationResolverParams, LocalComputedInputs } from '../utils'
-import { getPhotonDmmf } from './utils'
 import { DmmfDocument } from './DmmfDocument'
 import { DmmfTypes } from './DmmfTypes'
-import paginationStrategies, { PaginationStrategy } from '../pagination'
+import { getPhotonDmmf } from './utils'
 
 export type TransformOptions = {
   globallyComputedInputs?: GlobalComputedInputs
