+++ src/pagination/index.ts
@@ -1,0 +1,35 @@
+import { DMMF } from '@prisma/client/runtime'
+import { DmmfTypes } from '../dmmf'
+import { prismaStraegy } from './prisma'
+import { relayStrategy } from './relay'
+
+interface NormalizedPaginationArgs {
+  cursor?: object
+  skip?: string | number
+  take?: string | number
+  orderBy?: { [x: string]: 'asc' | 'desc' }
+  [x: string]: any
+}
+
+export interface PaginationStrategy<GraphQLPaginationArgs extends object = object> {
+  paginationArgNames: (keyof GraphQLPaginationArgs)[]
+  transformDmmfArgs(params: {
+    args: DmmfTypes.SchemaArg[]
+    paginationArgNames: string[]
+    field: DMMF.SchemaField
+  }): DmmfTypes.SchemaArg[]
+  resolve(args: GraphQLPaginationArgs): NormalizedPaginationArgs
+}
+
+// todo remove, but TS error if we do, dont' understand the error message
+interface PaginationStrategies {
+  relay: typeof relayStrategy
+  prisma: typeof prismaStraegy
+}
+
+export const paginationStrategies: PaginationStrategies = {
+  relay: relayStrategy,
+  prisma: prismaStraegy,
+}
+
+export type PaginationStrategyTypes = typeof relayStrategy | typeof prismaStraegy
